e001-01-wavelets-baseline/requirements.txt
torch
torchvision
torch-fidelity
wandb
pyyaml
matplotlib
